third_party/WebKit/LayoutTests/svg/custom/svg-parse-overflow-2-expected.txt

CONSOLE ERROR: line 9: Error: <rect> attribute height: A negative value is not valid. ("-105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
